backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart
pydantic
pyyaml
python-dotenv
httpx

# Computer Vision
opencv-python
Pillow

# Hugging Face & Transformers
transformers
accelerate
sentence-transformers

# FAISS - CPU version
faiss-cpu

# Audio processing
librosa
soundfile

# Document parsing
PyPDF2
python-docx

# Utilities
numpy
tqdm
colorama
imageio-ffmpeg

# PyTorch (install separately for CUDA support):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CPU-only:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
